Replace int type IDs with long

+++ PD.Domain/Entities/Order.cs
@@ -11,16 +11,16 @@ namespace PD.Domain.Entities
     public class Order
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public List<Pizza> Pizzas { get; set; }
         public string Adress { get; set; }
         public bool IsPaid { get; set; } = false;
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         [ForeignKey("UserId")]
         public User User { get; set; }
 
-        public int? PromoCodeId { get; set; }
+        public long? PromoCodeId { get; set; }
         [ForeignKey("PromoCodeId")]
         public PromoCode PromoCode { get; set; }
     }
