+++ PD.Web/Models/Order/OrderViewModel.cs
@@ -7,11 +7,11 @@ namespace PD.Web.Models
 {
     public class OrderViewModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public List<ShortPizzaViewModel> Pizzas { get; set; }
         public string Adress { get; set; }
         public bool IsPaid { get; set; }
-        public int UserId { get; set; }
-        public int? PromoCodeId { get; set; }
+        public long UserId { get; set; }
+        public long? PromoCodeId { get; set; }
     }
 }
